[#3558] OrderedGemCorrector extraction
